Make MySQL work.
